refactor(#123): api 디렉토리 구조 리팩토링
