+++ src/Layout/Navbar/Navbar.jsx
@@ -18,8 +18,8 @@ import LoginModal from "../../components/Modal/LoginModal/LoginModal";
 import SignUpModal from "../../components/Modal/SignUpModal/SignUpModal";
 import FindUserId from "../../components/Modal/FindUserIdModal/FindUserId";
 import FindUserPw from "../../components/Modal/FindUserPwModal/FindUserPw";
-import { logout } from "../../services/authApi";
-import { fetchCategoryItems } from "../../services/postApi";
+import { logout } from "../../services/api/authApi";
+import { fetchCategoryItems } from "../../services/api/postApi";
 import { userLogout } from "../../store/user/userSlice";
 import logoImg from "../../assets/images/DEEPDIVERS.png";
 import { selectMenuItem } from "../../store/category/menuSlice";
