+++ src/pages/Board/NoticeBoard/NoticeBoard.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { fetchNoticePostItems } from "../../../services/api";
+import { fetchNoticePostItems } from "../../../services/api/api";
 import NoticePostCard from "../../../components/Card/PostCard/NoticePostCard/NoticePostCard";
 import { BoardTitle, ContentWrapper, SearchSortWrapper } from "../Board.style";
 import Search from "../../../components/Common/Search/Search";
