+++ src/pages/MainPage/MainPage.jsx
@@ -17,7 +17,7 @@ import redirectIcon from "../../assets/images/redirect-to-board.png";
 import { useNavigate } from "react-router-dom";
 import projectBoardCardImg from "../../assets/images/Template Card.png";
 import freeBoardCardImg from "../../assets/images/Story Card Horizontal.png";
-import { fetchCategoryItems } from "../../services/postApi";
+import { fetchCategoryItems } from "../../services/api/postApi";
 import { useDispatch } from "react-redux";
 import { selectMenuItem } from "../../store/category/menuSlice";
