+++ src/services/api/postApi.js
@@ -1,4 +1,4 @@
-import axiosInstance from "./axiosConfig";
+import axiosInstance from "../axiosConfig";
 
 export const createPost = async (body, token) => {
   try {
